+++ backend/tests/baserow/contrib/database/field/test_field_handler.py
@@ -1,14 +1,17 @@
 import pytest
 from decimal import Decimal
+from unittest.mock import patch
 
 from baserow.core.exceptions import UserNotInGroupError
 from baserow.contrib.database.fields.handler import FieldHandler
 from baserow.contrib.database.fields.models import (
-    Field, TextField, NumberField, BooleanField
+    Field, TextField, NumberField, BooleanField, SelectOption
 )
+from baserow.contrib.database.fields.field_types import TextFieldType
+from baserow.contrib.database.fields.registries import field_type_registry
 from baserow.contrib.database.fields.exceptions import (
     FieldTypeDoesNotExist, PrimaryFieldAlreadyExists, CannotDeletePrimaryField,
-    FieldDoesNotExist, IncompatiblePrimaryFieldTypeError
+    FieldDoesNotExist, IncompatiblePrimaryFieldTypeError, CannotChangeFieldType
 )
 
 
@@ -47,14 +50,19 @@ def test_get_field(data_fixture):
 
 
 @pytest.mark.django_db
-def test_create_field(data_fixture):
+@patch('baserow.contrib.database.fields.signals.field_created.send')
+def test_create_field(send_mock, data_fixture):
     user = data_fixture.create_user()
     user_2 = data_fixture.create_user()
     table = data_fixture.create_database_table(user=user)
 
     handler = FieldHandler()
-    handler.create_field(user=user, table=table, type_name='text',
-                         name='Test text field', text_default='Some default')
+    field = handler.create_field(user=user, table=table, type_name='text',
+                                 name='Test text field', text_default='Some default')
+
+    send_mock.assert_called_once()
+    assert send_mock.call_args[1]['field'].id == field.id
+    assert send_mock.call_args[1]['user'].id == user.id
 
     assert Field.objects.all().count() == 1
     assert TextField.objects.all().count() == 1
@@ -143,11 +151,8 @@ def test_create_primary_field(data_fixture):
 
 
 @pytest.mark.django_db
-def test_update_field(data_fixture):
-    """
-    @TODO somehow trigger the CannotChangeFieldType and test if it is raised.
-    """
-
+@patch('baserow.contrib.database.fields.signals.field_updated.send')
+def test_update_field(send_mock, data_fixture):
     user = data_fixture.create_user()
     user_2 = data_fixture.create_user()
     table = data_fixture.create_database_table(user=user)
@@ -181,6 +186,10 @@ def test_update_field(data_fixture):
     assert field.name == 'Text field'
     assert field.text_default == 'Default value'
     assert isinstance(field, TextField)
+
+    send_mock.assert_called_once()
+    assert send_mock.call_args[1]['field'].id == field.id
+    assert send_mock.call_args[1]['user'].id == user.id
 
     # Insert some rows to the table which should be converted later.
     model = table.get_model()
@@ -238,7 +247,34 @@ def test_update_field(data_fixture):
 
 
 @pytest.mark.django_db
-def test_delete_field(data_fixture):
+def test_update_field_failing(data_fixture):
+    # This failing field type triggers the CannotChangeFieldType error if a field is
+    # changed into this type.
+    class FailingFieldType(TextFieldType):
+        def get_alter_column_type_function(self, connection, from_field, to_field):
+            return 'p_in::NOT_VALID_SQL_SO_IT_WILL_FAIL('
+
+    user = data_fixture.create_user()
+    table = data_fixture.create_database_table(user=user)
+    field = data_fixture.create_number_field(table=table, order=1)
+
+    handler = FieldHandler()
+
+    with patch.dict(
+        field_type_registry.registry,
+        {'text': FailingFieldType()}
+    ):
+        with pytest.raises(CannotChangeFieldType):
+            handler.update_field(user=user, field=field, new_type_name='text')
+
+    handler.update_field(user, field=field, new_type_name='text')
+    assert Field.objects.all().count() == 1
+    assert TextField.objects.all().count() == 1
+
+
+@pytest.mark.django_db
+@patch('baserow.contrib.database.fields.signals.field_deleted.send')
+def test_delete_field(send_mock, data_fixture):
     user = data_fixture.create_user()
     user_2 = data_fixture.create_user()
     table = data_fixture.create_database_table(user=user)
@@ -254,9 +290,15 @@ def test_delete_field(data_fixture):
 
     assert Field.objects.all().count() == 1
     assert TextField.objects.all().count() == 1
+    field_id = text_field.id
     handler.delete_field(user=user, field=text_field)
     assert Field.objects.all().count() == 0
     assert TextField.objects.all().count() == 0
+
+    send_mock.assert_called_once()
+    assert send_mock.call_args[1]['field_id'] == field_id
+    assert send_mock.call_args[1]['field'].id == field_id
+    assert send_mock.call_args[1]['user'].id == user.id
 
     table_model = table.get_model()
     field_name = f'field_{text_field.id}'
@@ -265,3 +307,97 @@ def test_delete_field(data_fixture):
     primary = data_fixture.create_text_field(table=table, primary=True)
     with pytest.raises(CannotDeletePrimaryField):
         handler.delete_field(user=user, field=primary)
+
+
+@pytest.mark.django_db
+def test_update_select_options(data_fixture):
+    user = data_fixture.create_user()
+    user_2 = data_fixture.create_user()
+    table = data_fixture.create_database_table(user=user)
+    field = data_fixture.create_single_select_field(table=table)
+    field_2 = data_fixture.create_single_select_field(table=table)
+    data_fixture.create_text_field(table=table)
+
+    handler = FieldHandler()
+
+    with pytest.raises(UserNotInGroupError):
+        handler.update_field_select_options(field=field, user=user_2, select_options=[])
+
+    handler.update_field_select_options(field=field, user=user, select_options=[
+        {'value': 'Option 1', 'color': 'blue'},
+        {'value': 'Option 2', 'color': 'red'}
+    ])
+
+    assert SelectOption.objects.all().count() == 2
+    select_options = field.select_options.all()
+    assert len(select_options) == 2
+
+    assert select_options[0].order == 0
+    assert select_options[0].value == 'Option 1'
+    assert select_options[0].color == 'blue'
+    assert select_options[0].field_id == field.id
+
+    assert select_options[1].order == 1
+    assert select_options[1].value == 'Option 2'
+    assert select_options[1].color == 'red'
+    assert select_options[1].field_id == field.id
+
+    handler.update_field_select_options(field=field, user=user, select_options=[
+        {'id': select_options[0].id, 'value': 'Option 1 A', 'color': 'blue 2'},
+        {'id': select_options[1].id, 'value': 'Option 2 A', 'color': 'red 2'}
+    ])
+
+    assert SelectOption.objects.all().count() == 2
+    select_options_2 = field.select_options.all()
+    assert len(select_options_2) == 2
+
+    assert select_options_2[0].id == select_options[0].id
+    assert select_options_2[0].order == 0
+    assert select_options_2[0].value == 'Option 1 A'
+    assert select_options_2[0].color == 'blue 2'
+    assert select_options_2[0].field_id == field.id
+
+    assert select_options_2[1].id == select_options[1].id
+    assert select_options_2[1].order == 1
+    assert select_options_2[1].value == 'Option 2 A'
+    assert select_options_2[1].color == 'red 2'
+    assert select_options_2[1].field_id == field.id
+
+    handler.update_field_select_options(field=field, user=user, select_options=[
+        {'id': select_options[1].id, 'value': 'Option 1 B', 'color': 'red'},
+        {'value': 'Option 2 B', 'color': 'green'},
+    ])
+
+    assert SelectOption.objects.all().count() == 2
+    select_options_3 = field.select_options.all()
+    assert len(select_options_3) == 2
+
+    assert select_options_3[0].id == select_options[1].id
+    assert select_options_3[0].order == 0
+    assert select_options_3[0].value == 'Option 1 B'
+    assert select_options_3[0].color == 'red'
+    assert select_options_3[0].field_id == field.id
+
+    assert select_options_3[1].order == 1
+    assert select_options_3[1].value == 'Option 2 B'
+    assert select_options_3[1].color == 'green'
+    assert select_options_3[1].field_id == field.id
+
+    handler.update_field_select_options(field=field_2, user=user, select_options=[
+        {'id': select_options[1].id, 'value': 'Option 1 B', 'color': 'red'},
+    ])
+
+    assert SelectOption.objects.all().count() == 3
+    select_options_4 = field_2.select_options.all()
+    assert len(select_options_4) == 1
+
+    assert select_options_4[0].id != select_options[1].id
+    assert select_options_4[0].order == 0
+    assert select_options_4[0].value == 'Option 1 B'
+    assert select_options_4[0].color == 'red'
+    assert select_options_4[0].field_id == field_2.id
+
+    handler.update_field_select_options(field=field_2, user=user, select_options=[])
+
+    assert SelectOption.objects.all().count() == 2
+    assert field_2.select_options.all().count() == 0