+++ web-frontend/modules/database/plugin.js
@@ -10,6 +10,7 @@ import {
   BooleanFieldType,
   DateFieldType,
   FileFieldType,
+  SingleSelectFieldType,
 } from '@baserow/modules/database/fieldTypes'
 import {
   EqualViewFilterType,
@@ -20,6 +21,8 @@ import {
   ContainsNotViewFilterType,
   HigherThanViewFilterType,
   LowerThanViewFilterType,
+  SingleSelectEqualViewFilterType,
+  SingleSelectNotEqualViewFilterType,
   BooleanViewFilterType,
   EmptyViewFilterType,
   NotEmptyViewFilterType,
@@ -34,6 +37,8 @@ import tableStore from '@baserow/modules/database/store/table'
 import viewStore from '@baserow/modules/database/store/view'
 import fieldStore from '@baserow/modules/database/store/field'
 import gridStore from '@baserow/modules/database/store/view/grid'
+
+import { registerRealtimeEvents } from '@baserow/modules/database/realtime'
 
 export default ({ store, app }) => {
   store.registerModule('table', tableStore)
@@ -51,6 +56,8 @@ export default ({ store, app }) => {
   app.$registry.register('viewFilter', new ContainsNotViewFilterType())
   app.$registry.register('viewFilter', new HigherThanViewFilterType())
   app.$registry.register('viewFilter', new LowerThanViewFilterType())
+  app.$registry.register('viewFilter', new SingleSelectEqualViewFilterType())
+  app.$registry.register('viewFilter', new SingleSelectNotEqualViewFilterType())
   app.$registry.register('viewFilter', new BooleanViewFilterType())
   app.$registry.register('viewFilter', new EmptyViewFilterType())
   app.$registry.register('viewFilter', new NotEmptyViewFilterType())
@@ -63,7 +70,10 @@ export default ({ store, app }) => {
   app.$registry.register('field', new URLFieldType())
   app.$registry.register('field', new EmailFieldType())
   app.$registry.register('field', new FileFieldType())
+  app.$registry.register('field', new SingleSelectFieldType())
   app.$registry.register('importer', new CSVImporterType())
   app.$registry.register('importer', new PasteImporterType())
   app.$registry.register('settings', new APITokenSettingsType())
+
+  registerRealtimeEvents(app.$realtime)
 }